Refactor move Environment DTO from IMPL to API module
